Furnished university basement (#9)
* Increased basement height

Increased the height of the university's basement by deleting the terrain in the area the university occupies and then moving the floor downward and extending the walls.

* Added initial furnishings to the library

Added bookshelves and reading tables in the basement of the university

* Added books in bookshelves in library

Added books in all the bookshelves in the library. Finished furnishing the library.